--- scripts/install/.colcon_install_layout
@@ -1,1 +1,0 @@
-isolated
